AshesOfTomorrow/Windfarm/requirements.txt
Flask==1.0.3
pymongo==3.7.2
pycryptodomex==3.8.2
pycrypto==2.6.1
dnspython==1.16.0
gunicorn

